requirements.txt
# Selenium and WebDriver
selenium==4.15.2
webdriver-manager==4.0.1

# Testing Framework
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0

# Test Data Generation
Faker==20.1.0

# Reporting
allure-pytest==2.13.2

# Configuration
python-dotenv==1.0.0

# Utilities
Pillow==10.1.0
